apps/agent/pyproject.toml
[project]
name = "voicecraft-agent"
version = "0.1.0"
description = "VoiceCraft LiveKit voice agent worker"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.4.0",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-google>=1.0.0",
    "livekit-plugins-elevenlabs>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "deepgram-sdk>=3.0.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
